Versión final pequeños arreglos

+++ src/main/java/es/udc/tfgproject/backend/rest/controllers/commercialMedicamentController.java
@@ -15,19 +15,27 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import es.udc.tfgproject.backend.model.services.ListService;
+import es.udc.tfgproject.backend.model.services.SecService;
 import es.udc.tfgproject.backend.rest.dtos.CommercialMedicamentDto;
 import es.udc.tfgproject.backend.rest.dtos.MedicamentExtendedDto;
 
 @Transactional
 @ControllerAdvice
-@RequestMapping("/administracion/medicamentosComerciales")
+@RequestMapping("/administracion/medicamentos")
 public class commercialMedicamentController {
 
     @Autowired
     private ListService listService;
+    @Autowired
+    private SecService secService;
 
     @GetMapping("")
     public String commercialMedicaments(Model model) {
+	boolean security = secService.checkSecurity("ROLE_ADMIN");
+
+	if (!security) {
+	    return "forbidden";
+	}
 	ArrayList<CommercialMedicamentDto> commercialMedicamentsList = listService.listAllCommercialMedicamentDto();
 
 	model.addAttribute("medicaments", commercialMedicamentsList);
@@ -37,6 +45,11 @@ public class commercialMedicamentController {
 
     @GetMapping("/eliminar/{code}")
     public String deleteCommercialMedicament(@PathVariable("code") String code, Model model) {
+	boolean security = secService.checkSecurity("ROLE_ADMIN");
+
+	if (!security) {
+	    return "forbidden";
+	}
 	ArrayList<CommercialMedicamentDto> commercialMedicamentsList = listService.listAllCommercialMedicamentDto();
 
 	listService.deleteCommercialMedicamentByCode(commercialMedicamentsList, code);
@@ -53,14 +66,20 @@ public class commercialMedicamentController {
     @GetMapping("/editar/{code}")
     public String editCommercialMedicament(@PathVariable("code") String code, Model model) {
 
+	boolean security = secService.checkSecurity("ROLE_ADMIN");
+
+	if (!security) {
+	    return "forbidden";
+	}
+
 	CommercialMedicamentDto medicament = listService.getCommercialMedicamentByCode(code);
 
 	if (medicament.getCommercialMedicamentName() == null || medicament.getCommercialMedicamentName() == "") {
-	    model.addAttribute("title", "Nuevo Medicamento comercial");
-	    model.addAttribute("err", "No se ha encontrado el medicamento comercial, si guarda se insertará una nuevo");
+	    model.addAttribute("title", "Nuevo Medicamento");
+	    model.addAttribute("err", "No se ha encontrado el medicamento, si guarda se insertará una nuevo");
 	    model.addAttribute("medicament", medicament);
 	} else {
-	    model.addAttribute("title", "Detalle Medicamento comercial");
+	    model.addAttribute("title", "Detalle Medicamento");
 	    model.addAttribute("medicament", medicament);
 	    model.addAttribute("oldCommercialMedicamentName", medicament.getCommercialMedicamentName());
 	}
@@ -74,8 +93,13 @@ public class commercialMedicamentController {
 
     @GetMapping("/nuevo")
     public String newCommercialMedicament(Model model) {
+	boolean security = secService.checkSecurity("ROLE_ADMIN");
 
-	model.addAttribute("title", "Nuevo Medicamento comercial");
+	if (!security) {
+	    return "forbidden";
+	}
+
+	model.addAttribute("title", "Nuevo Medicamento");
 
 	CommercialMedicamentDto medicament = new CommercialMedicamentDto();
 	model.addAttribute("medicament", medicament);
@@ -92,15 +116,21 @@ public class commercialMedicamentController {
 	    @RequestParam(value = "oldCommercialMedicamentName", required = false) String oldCommercialMedicamentName,
 	    @RequestParam(value = "medicamentName", required = false) String medicamentName, Model model) {
 
+	boolean security = secService.checkSecurity("ROLE_ADMIN");
+
+	if (!security) {
+	    return "forbidden";
+	}
+
 	Boolean hasError = listService.checkAndSaveCommercialMedicament(oldCommercialMedicamentName,
 		medicament.getCommercialMedicamentName(), medicamentName);
 
 	if (hasError) {
-	    model.addAttribute("err", "Ya existe un medicamento comercial con ese nombre");
+	    model.addAttribute("err", "Ya existe un medicamento con ese nombre");
 	    if (oldCommercialMedicamentName == "") {
-		model.addAttribute("title", "Nuevo Medicamento comercial");
+		model.addAttribute("title", "Nuevo Medicamento");
 	    } else {
-		model.addAttribute("title", "Detalle Medicamento comercial");
+		model.addAttribute("title", "Detalle Medicamento");
 		model.addAttribute("medicament", medicament);
 		model.addAttribute("oldCommercialMedicamentName", oldCommercialMedicamentName);
 	    }
